String: remove innards from public header.
A dom_string is the first member of the underlying
dom_string_internal and thus their addresses are equivalent. We
shouldn't be exposing the internal alignment requirements through
the public header at all, so remove the bogus _ALIGNED and cast
through void internally to tell the compiler that we know what
we're doing.